% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_subfunctions.R
\name{ask_new_condaenv_name}
\alias{ask_new_condaenv_name}
\title{Subfunction to ask new conda environment name.}
\usage{
ask_new_condaenv_name(env_name = NULL)
}
\arguments{
\item{env_name}{Environment name (default to NULL).}
}
\value{
env_name  Environment name entered by the user or same name as the input.
}
\description{
This function asks the user to enter a name when input variable
(env_name) is not specified.
When input env_name is specified, the same env_name is returned
if the same named conda environment does not exist.
}
