% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_permission.R
\name{ask_permission}
\alias{ask_permission}
\title{Subfunction to ask user's permission
for the query.}
\usage{
ask_permission(query)
}
\arguments{
\item{query}{Query}
}
\value{
Boolean
}
\description{
Subfunction to ask user's permission
for the query.
}
