% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_to_existing_pythonenv.R
\name{choose_python_binary}
\alias{choose_python_binary}
\title{Subfunction to choose a Python binary.}
\usage{
choose_python_binary()
}
\value{
python  File path of the chosen Python binary.
}
\description{
This function asks user to choose a Python binary.
Local Python binaries in AppData\\Local\\Programs\\Python or ~/.local/bin are
automatically found and suggested as options.
}
