% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_subfunctions.R
\name{create_new_recode_condaenv}
\alias{create_new_recode_condaenv}
\title{Subfunction to create a new RECODE-enabled conda environment.}
\usage{
create_new_recode_condaenv(env_name = NULL, if_test = FALSE)
}
\arguments{
\item{env_name}{New environment name, default to NULL.}

\item{if_test}{Flag for unit test mode.}
}
\value{
env_name  Created conda environment name.
}
\description{
This function creates a new conda environment with reticulate::conda_create,
then installs RECODE to the created environment with reticulate::conda_install.
}
