% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_local_python_binary.R
\name{find_local_python_binary}
\alias{find_local_python_binary}
\title{Subfunction to find local system Python binaries.}
\usage{
find_local_python_binary()
}
\value{
pythons  List of found Python binaries.
}
\description{
This function finds local Python binaries depending on OS.
}
