% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_local_python_binary.R
\name{find_python_in_AppData}
\alias{find_python_in_AppData}
\title{Subfunction to find Python binary in AppData\\Local\\Programs\\Python.}
\usage{
find_python_in_AppData()
}
\value{
appdata_pythons  Python binaries in AppData\\Local\\Programs\\Python.
}
\description{
This function finds local Python in Windows.
}
