% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_if_python_binary.R
\name{guess_if_python_binary}
\alias{guess_if_python_binary}
\title{Subfunction to guess if a specified file is a Python binary}
\usage{
guess_if_python_binary(file)
}
\arguments{
\item{file}{File to check}
}
\value{
Boolean
}
\description{
This function checks
\itemize{
\item file existence
\item if filename is consistent with Python binary
}
}
