% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_subfunctions.R
\name{if_condaenvname_already_exists}
\alias{if_condaenvname_already_exists}
\title{Subfunction to check if the same conda environment
as the input name already exists.}
\usage{
if_condaenvname_already_exists(env_name)
}
\arguments{
\item{env_name}{Input environment name to check.}
}
\value{
Boolean
}
\description{
Subfunction to check if the same conda environment
as the input name already exists.
}
