% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_to_existing_virtualenv.R
\name{if_virtualenv_exists}
\alias{if_virtualenv_exists}
\title{Subfunction to check if virtualenv already exists.}
\usage{
if_virtualenv_exists()
}
\value{
Boolean
}
\description{
This function uses reticulate::virtualenv_list.
}
