% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_new_miniconda.R
\name{install_new_miniconda}
\alias{install_new_miniconda}
\title{Install Miniconda to create a new RECODE-enabled conda environment.}
\usage{
install_new_miniconda()
}
\value{
install_info  Vector of install information.
}
\description{
This function installs Miniconda with reticulate::install_minicona,
and create a new conda environment to install RECODE.
}
