% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_screcode.R
\name{install_screcode}
\alias{install_screcode}
\title{Install the Python version of RECODE (screcode) for R.}
\usage{
install_screcode()
}
\description{
This function enables the user to install
the Python version RECODE by using the R reticulate package.\cr
\cr
RECODE - resolution of curse of dimensionality in single-cell data analysis : https://github.com/yusuke-imoto-lab/RECODE \cr
\cr
Ushey K, Allaire J, Tang Y (2023). reticulate: Interface to 'Python'. https://rstudio.github.io/reticulate/, https://github.com/rstudio/reticulate. \cr
}
\details{
This function supports the following types of RECODE installation.
\itemize{
\item Install Miniconda to create a new RECODE-enabled conda environment (if conda is not installed, recommended)..
\item Install RECODE to a new or existing conda environment (if conda is installed, recommended).
\item Install RECODE to an existing virtualenv environment.
\item Install RECODE using a specified Python binary (e.g., local Python, Python embeddable).
}

The user can choose their preferred installation type
according to the guidance.

After the installation is done, a directory called "recodeloader",
is created in the working directory.
This directory contains an R script for loading the RECODE-enabled Python.

The user can load the created RECODE-enabled Python by running

source("recodeloader/load_recodeenv.R")

in the working directory.
}
