% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_to_another_pythonenv.R
\name{install_to_another_pythonenv}
\alias{install_to_another_pythonenv}
\title{Install RECODE to another Python environment other than conda.}
\usage{
install_to_another_pythonenv()
}
\value{
install_info  Vector of install information.
}
\description{
This function enables the user to install RECODE to
an existing virtualenv or a Python environment using
a specified binary.
}
