% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_to_condaenv.R
\name{install_to_condaenv}
\alias{install_to_condaenv}
\title{Install RECODE to new or existing conda environment.}
\usage{
install_to_condaenv()
}
\value{
install_info  Vector of install information.
}
\description{
This function creates a new RECODE-enabled conda environment
or install RECODE to an existing conda environment.
The reticulate::conda_create and reticulate::conda_install
functions are used.
}
