% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_subfunctions.R
\name{install_to_existing_condaenv}
\alias{install_to_existing_condaenv}
\title{Subfunction to install RECODE to an existing conda environment.}
\usage{
install_to_existing_condaenv(env_name, if_test = FALSE)
}
\arguments{
\item{env_name}{Environment name}

\item{if_test}{Flag for unit test mode.}
}
\value{
Name of environment to which RECODE is installed.
}
\description{
This function uses reticulate::conda_install.
}
