% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_to_existing_pythonenv.R
\name{install_to_existing_pythonenv}
\alias{install_to_existing_pythonenv}
\title{Install RECODE using a specified Python binary.}
\usage{
install_to_existing_pythonenv()
}
\value{
python  File path of the chosen Python binary.
}
\description{
Install RECODE using a specified Python binary.
}
