% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_to_existing_virtualenv.R
\name{install_to_existing_virtualenv}
\alias{install_to_existing_virtualenv}
\title{Install RECODE to an existing virtualenv.}
\usage{
install_to_existing_virtualenv(installation_choice = NULL)
}
\arguments{
\item{installation_choice}{Default to NULL, specified iff test mode.}
}
\value{
virtualenv name to which RECODE is installed.
}
\description{
This function installs RECODE to an existing virtualenv environment
with reticulate::virtualenv_install.
Note that creation of new virtualenv is not supported.
}
