% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_subfunctions.R
\name{is_conda_command_available}
\alias{is_conda_command_available}
\title{Subfunction to check if conda command is available.}
\usage{
is_conda_command_available()
}
\value{
Boolean
}
\description{
The reticulate::conda_list function is used.
}
