% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_if_python_binary.R
\name{is_filename_python}
\alias{is_filename_python}
\title{Subfunction to check if a specified file
has Python binary file name such as  "python.exe" or "python.d{1}.d{1,2}".}
\usage{
is_filename_python(file)
}
\arguments{
\item{file}{File to check.}
}
\value{
Boolean
}
\description{
Subfunction to check if a specified file
has Python binary file name such as  "python.exe" or "python.d{1}.d{1,2}".
}
