% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_loader.R
\name{make_loader}
\alias{make_loader}
\title{Make an R script which loads RECODE-enabled Python environment.}
\usage{
make_loader(install_info)
}
\arguments{
\item{install_info}{Install information}
}
\value{
File path of created "load_recodeenv.R" is returned.
}
\description{
This function creates directory "recodeloader" in
the working directory to output an R script "load_recodeenv.R"
into the directory.
If installation is done using a Python binary,
.RData file "install_info.RData",
which contains Python's path information,
is also created in the same directory.
}
\details{
The user can load the created RECODE-enabled Python by running

source("recodeloader/load_recodeenv.R")

in the working directory.
}
